fix : 필드 검증 수정, short -> Int 변경 및 HistoryEntity 생성 방식 변경

+++ src/main/kotlin/com/Dnight/calinify/event/dto/response/EventResponseDTO.kt
@@ -13,7 +13,7 @@ class EventResponseDTO(
     val description : String? = null,
     val startAt : LocalDateTime,
     val endAt : LocalDateTime,
-    val priority : Short?,
+    val priority : Int?,
     val location : String? = null,
     val repeatRule : String? = null,
     val status : EventStatus,
